fix / avoid songs with no lyrics
Instead of taking the first match of a song and print its lyrics right
away, we know iterate over the list of songs returned and skip the ones
which lyrics are equal to '(lyrics not available)'. If all the songs'
lyrics happen to be equal that message then the last song of the list is
printed.